backend/requirements.txt
fastapi
uvicorn[standard]
httpx
qdrant-client>=1.6.0
requests
sentence_transformers
langchain
pymupdf
pytesseract
pillow
keybert
langdetect
python-docx
python-multipart
debugpy
chonkie
SQLAlchemy
psycopg2-binary
trafilatura
beautifulsoup4
lxml_html_clean
python-jose[cryptography]
authlib
pydantic[email]
pydantic-settings